CLI-839 Move wrapText next to its only caller and drop messages.ts
